change package names to org.jboss.set to reflect github space change
