--- src/main/java/io/xstefank/exception/PKBIndexNotLoadedException.java
@@ -1,7 +1,0 @@
-package io.xstefank.exception;
-
-public class PKBIndexNotLoadedException extends Exception {
-    public PKBIndexNotLoadedException(Exception e) {
-        super(e);
-    }
-}
