--- src/main/java/io/xstefank/model/json/BuildModifyInfo.java
@@ -1,30 +1,0 @@
-package io.xstefank.model.json;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-
-public class BuildModifyInfo {
-
-    @NotNull
-    public String project;
-
-    @NotNull
-    @Valid
-    public Environment environment;
-
-    @NotNull
-    public String upstream;
-
-    @NotNull
-    public String script;
-
-    @Override
-    public String toString() {
-        return "BuildModifyInfo{" +
-            "project='" + project + '\'' +
-            ", environment=" + environment +
-            ", upstream='" + upstream + '\'' +
-            ", script='" + script + '\'' +
-            '}';
-    }
-}
