--- src/main/java/io/xstefank/model/json/Environment.java
@@ -1,21 +1,0 @@
-package io.xstefank.model.json;
-
-import jakarta.validation.constraints.Size;
-import java.util.List;
-
-public class Environment {
-
-    @Size(min = 1)
-    public List<String> openjdk;
-
-    @Size(min = 1)
-    public List<String> maven;
-
-    @Override
-    public String toString() {
-        return "Environment{" +
-            "openjdk=" + openjdk +
-            ", maven=" + maven +
-            '}';
-    }
-}
