--- src/main/java/io/xstefank/pkb/model/json/EnvironmentSpec.java
@@ -1,9 +1,0 @@
-package io.xstefank.pkb.model.json;
-
-import java.util.List;
-
-public class EnvironmentSpec {
-
-    public Tools tools;
-    public List<String> requiredTools;
-}
