--- src/main/java/io/xstefank/pkb/model/json/LastBuild.java
@@ -1,7 +1,0 @@
-package io.xstefank.pkb.model.json;
-
-public class LastBuild {
-
-    public String upstreamVersion;
-    public String prodVersion;
-}
