--- src/main/java/io/xstefank/pkb/model/json/Source.java
@@ -1,7 +1,0 @@
-package io.xstefank.pkb.model.json;
-
-public class Source {
-
-    public String internalScmUrl;
-    public String scmUrl;
-}
