--- src/main/java/io/xstefank/pkb/model/json/Tools.java
@@ -1,7 +1,0 @@
-package io.xstefank.pkb.model.json;
-
-public class Tools {
-
-    public String openjdk;
-    public String maven;
-}
