+++ src/main/java/org/jboss/set/exception/PKBIndexNotLoadedException.java
@@ -1,0 +1,7 @@
+package org.jboss.set.exception;
+
+public class PKBIndexNotLoadedException extends Exception {
+    public PKBIndexNotLoadedException(Exception e) {
+        super(e);
+    }
+}
