+++ src/main/java/org/jboss/set/pkb/model/json/BuildConfiguration.java
@@ -1,0 +1,17 @@
+package org.jboss.set.pkb.model.json;
+
+import java.util.Map;
+
+public class BuildConfiguration {
+
+    public String id;
+    public String environment;
+    public EnvironmentSpec environmentSpec;
+    public String imageId;
+    public String versionPattern;
+    public String script;
+    public String testScriptVars;
+    public Source source;
+    public String system;
+    public Map<String, String> parameters;
+}
