+++ src/main/java/org/jboss/set/pkb/model/json/EnvironmentSpec.java
@@ -1,0 +1,9 @@
+package org.jboss.set.pkb.model.json;
+
+import java.util.List;
+
+public class EnvironmentSpec {
+
+    public Tools tools;
+    public List<String> requiredTools;
+}
