+++ src/main/java/org/jboss/set/pkb/model/json/LastBuild.java
@@ -1,0 +1,7 @@
+package org.jboss.set.pkb.model.json;
+
+public class LastBuild {
+
+    public String upstreamVersion;
+    public String prodVersion;
+}
