+++ src/main/java/org/jboss/set/pkb/model/json/Project.java
@@ -1,0 +1,27 @@
+package org.jboss.set.pkb.model.json;
+
+import java.util.List;
+
+public class Project {
+
+    public String project;
+    public String owner;
+    public String stream;
+    public String ga;
+    public BuildConfiguration buildConfiguration;
+    public LastBuild lastBuild;
+    public List<String> products;
+
+    @Override
+    public String toString() {
+        return "Project{" +
+            "project='" + project + '\'' +
+            ", owner='" + owner + '\'' +
+            ", stream='" + stream + '\'' +
+            ", ga='" + ga + '\'' +
+            ", buildConfiguration=" + buildConfiguration +
+            ", lastBuild=" + lastBuild +
+            ", products=" + products +
+            '}';
+    }
+}
