+++ src/main/java/org/jboss/set/pkb/model/json/Source.java
@@ -1,0 +1,7 @@
+package org.jboss.set.pkb.model.json;
+
+public class Source {
+
+    public String internalScmUrl;
+    public String scmUrl;
+}
