+++ src/main/java/org/jboss/set/pkb/model/json/Tools.java
@@ -1,0 +1,7 @@
+package org.jboss.set.pkb.model.json;
+
+public class Tools {
+
+    public String openjdk;
+    public String maven;
+}
