--- src/test/java/io/xstefank/NativeBuildTriggerResourceIT.java
@@ -1,9 +1,0 @@
-package io.xstefank;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class NativeBuildTriggerResourceIT extends BuildTriggerResourceTest {
-
-    // Execute the same tests but in native mode.
-}
