+++ src/test/java/org/jboss/set/NativeBuildTriggerResourceIT.java
@@ -1,0 +1,9 @@
+package org.jboss.set;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class NativeBuildTriggerResourceIT extends BuildTriggerResourceTest {
+
+    // Execute the same tests but in native mode.
+}
